Search for .env files in possible project root directories
Storing an .env file in the project root (i.e. alongside manage.py)
would be a common use case. We can try and infer the location of the
project root based on how the application was invoked and by
analysing sys.argv.

Note that since we already check execution paths and their parent
directories, this additional search does not improve the ability
to find a .env file in the project root if the venv is also located
there.