Ignore type check for __file__ attribute on dynamically loaded modules
mypy throws the following error when the __file__ attribute of a dynamically
loaded module is converted to a pathlib.Path:

```
Argument 1 to "Path" has incompatible type "Optional[str]"; expected "Union[str, PathLike[str]]"
```

It looks like there are better methods than using the __file__ attribute for
retrieving the same information (e.g. importlib.resources for Python 3.7+). But
continuing to use this method for now shouldn't be an issue based on the
control flow (i.e. ModuleNotFoundError is raised before the __file__ attribute
is accessed for non-existent modules) and the use cases (i.e. loading a module
which was generated by Django's startproject).